первичная реализация skeletom loader, оптимизация преобразования записей

+++ app/src/main/java/com/example/movie_rec_sys/data/FirestoreRepository.kt
@@ -43,7 +43,7 @@ class FirestoreRepository(
                             marked = hash["marked"] as Boolean,
                             viewed = hash["viewed"] as Boolean,
                             rated = hash["rated"] as? Int,
-                            relevanceIndex = hash["relevance_index"] as Int,
+                            relevanceIndex = (hash["relevance_index"] as Long).toInt(),
                         )
                         Triple(actionFlag, docId, docContent)
                     }
